Move libraries into lib folder, organize Material classes
