+++ src/CleanTF2.CLI/lib/license_VTFCmd.txt
