+++ src/CleanTF2.CLI/lib/license_VTFLib.txt
