+++ src/CleanTF2.Core.Tests/TheFlatMaterialGenerator.cs
@@ -1,4 +1,5 @@
-﻿using CleanTF2.Core.Utilities;
+﻿using CleanTF2.Core.Materials;
+using CleanTF2.Core.Utilities;
 
 namespace CleanTF2.Core.Tests
 {
