+++ src/CleanTF2.Core/Materials/MaterialExtractor.cs
@@ -1,7 +1,7 @@
 ﻿using CleanTF2.Core.Utilities;
 using CleanTF2.Core.Valve;
 
-namespace CleanTF2.Core
+namespace CleanTF2.Core.Materials
 {
     public class MaterialExtractor : IMaterialExtractor
     {
